Test folder cleanup to improve unit test output
