--- test/unittest_transport.cpp
@@ -1,26 +1,0 @@
-#include "unittest_transport.h"
-/*
-#include "mbed.h"
-
-#define PIN_SWD_TX    PA_9
-#define PIN_SWD_RX    PA_10
-
-Serial serial(PIN_SWD_TX, PIN_SWD_RX, "serial");
-
-void unittest_uart_begin()
-{
-    serial.baud(115200);
-    serial.printf("\nSerial interface started...\n");
-    freopen("/serial", "w", stdout);  // retarget stdout
-}
-
-void unittest_uart_putchar(int c)
-{
-    serial.putc(c);
-}
-
-void unittest_uart_flush(){}
-
-void unittest_uart_end() {
-}
-*/
