--- test/unittest_transport.h
@@ -1,17 +1,0 @@
-#ifndef UNITEST_TRANSPORT_H
-#define UNITEST_TRANSPORT_H
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-void unittest_uart_begin();
-void unittest_uart_putchar(int c);
-void unittest_uart_flush();
-void unittest_uart_end();
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif // UNITEST_TRANSPORT_H
